pyproject.toml
[project]
name = "llm-test"
version = "0.1.0"
description = "A simple command-line application for testing interactions with Anthropic's Claude AI models"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
llm-test = "llm_test:main"
